type: move all database interface to a dbmodel type file
