+++ src/models/inspection.model.ts
@@ -1,24 +1,11 @@
-import { Schema, model, Document } from 'mongoose';
-import { ProductDoc } from './product.model';
-import { UserDoc } from './user.model';
-
+import { Schema, model } from 'mongoose';
+import { InspectionDoc } from '../types/dbmodel';
 // Define enum for functionality status
 enum FunctionalityStatus {
   Working = 'working',
   NotWorking = 'not_working',
 }
 
-// Define interface for Inspection document
-interface InspectionDoc extends Document {
-  product_id: ProductDoc['_id'];
-  inspector_id: UserDoc['_id'];
-  inspection_date: Date;
-  functionality_status: FunctionalityStatus;
-  additional_information?: string | null;
-  images?: string[] | null;
-  created_at: Date;
-  updated_at: Date;
-}
 
 // Define Inspection Schema
 const InspectionSchema = new Schema<InspectionDoc>({
@@ -35,4 +22,4 @@ const InspectionSchema = new Schema<InspectionDoc>({
 // Define and export Inspection model
 const Inspection = model<InspectionDoc>('Inspection', InspectionSchema);
 
-export { Inspection, FunctionalityStatus, InspectionDoc };
+export { Inspection, FunctionalityStatus};
