+++ src/types/dbmodel.ts
@@ -1,5 +1,13 @@
 import { Types } from 'mongoose';
 import { UserRole } from '../models/user.model';
+import { TrackerStatus } from '../models/supplier_inventory.model';
+import { AddressType } from '../models/address.model';
+import { GadgetType,  StateOfProduct, MemoryType } from '../models/product.model';
+import { OrderStatus } from '../models/order.model';
+import { DeliveryStatus, ModeOfDelivery } from '../models/logistics_delivery.model';
+import { InventoryStatus } from '../models/inventory.model';
+import { FunctionalityStatus } from '../models/inspection.model';
+
 
 // Define interface for Permission document
 interface PermissionDoc extends Document {
@@ -40,4 +48,136 @@ interface SupplierDoc extends Document {
   updated_at: Date;
 }
 
-export { PermissionDoc, UserDoc, SupplierDoc };
+// Define interface for SupplierInventory document
+interface SupplierInventoryDoc extends Document {
+  product_id: Types.ObjectId
+  supplier_id: Types.ObjectId;
+  quantity?: number;
+  tracker_status: TrackerStatus;
+  created_at: Date;
+  updated_at: Date;
+}
+
+
+// Define interface for Product document
+interface ProductDoc extends Document {
+  supplier_id: Types.ObjectId;
+  created_by: Types.ObjectId;
+  created_at: Date;
+  arrived_at?: Date | null;
+  sold_at?: Date | null;
+  updated_at: Date;
+  gadget_type: GadgetType;
+  brand: string;
+  gadget_model: string;
+  serial_number: string | null;
+  ram?: number | null;
+  rom?: number | null;
+  is_arrived: boolean;
+  returned_at?: Date | null;
+  memory_type?: MemoryType | null;
+  os?: string | null;
+  screen_size?: number | null;
+  price: number;
+  quantity: number;
+  state_of_product: StateOfProduct;
+  processor?: string | null;
+  graphics_card?: string | null;
+  backlit_keyboard: boolean;
+  touchscreen: boolean;
+  convertible: boolean;
+  fingerprint_scanner: boolean;
+  face_unlock: boolean;
+  camera_megapixels?: number | null;
+  sim_type?: string | null;
+  network_connectivity?: string | null;
+  waterproof: boolean;
+  wireless_charging: boolean;
+  stylus_included: boolean;
+  cellular_connectivity: boolean;
+  other_features?: string[] | null;
+}
+
+
+// Define interface for PaymentMethod document
+interface PaymentMethodDoc extends Document {
+  user_id: Types.ObjectId;
+  account_name: string;
+  account_number: number;
+  bank_name: string;
+  created_at: Date;
+  updated_at: Date;
+}
+
+// Define interface for Order document
+interface OrderDoc extends Document {
+  user_id: Types.ObjectId;
+  customer_name: string;
+  status: OrderStatus;
+  total_amount: number | null;
+  created_at: Date;
+  updated_at: Date;
+}
+
+// Define interface for OrderItem document
+interface OrderItemDoc extends Document {
+  order_id: Types.ObjectId;
+  product_id: Types.ObjectId;
+  quantity: number;
+  discount_amount?: number | null;
+  discount_percentage?: number | null;
+  unit_price: number | null;
+  created_at: Date;
+  updated_at: Date;
+}
+
+// Define interface for LogisticsDelivery document
+interface LogisticsDeliveryDoc extends Document {
+  order_id: Types.ObjectId;
+  user_id: Types.ObjectId;
+  courier_provider_name: string;
+  tracking_number?: string | null;
+  estimated_delivery_date: Date;
+  delivery_status: DeliveryStatus;
+  mode_of_delivery?: ModeOfDelivery;
+  delivery_fee?: number | null;
+  created_at: Date;
+  updated_at: Date;
+}
+
+// Define interface for Inventory document
+interface InventoryDoc extends Document {
+  quantity_in_stock: number;
+  inventory_status: InventoryStatus;
+  product_id:  Types.ObjectId;
+  location_in_warehouse?: string;
+  supplier_id:  Types.ObjectId;
+  created_at: Date;
+  updated_at: Date;
+}
+
+// Define interface for Inspection document
+interface InspectionDocument extends Document {
+  quantity_in_stock: number;
+  inventory_status: InventoryStatus;
+  product_id:  Types.ObjectId;
+  location_in_warehouse?: string;
+  supplier_id:  Types.ObjectId;
+  created_at: Date;
+  updated_at: Date;
+}
+
+
+
+// interface InspectionDoc extends Document {
+//   product_id: Types.ObjectId;
+//   inspector_id: Types.ObjectId;
+//   inspection_date: Date;
+//   functionality_status: FunctionalityStatus;
+//   additional_information?: string | null;
+//   images: HTMLCollectionOf<HTMLImageElement> | null;
+//   created_at: Date;
+//   updated_at: Date;
+// }
+
+export { PermissionDoc, UserDoc, SupplierDoc, SupplierInventoryDoc, ProductDoc, PaymentMethodDoc, OrderDoc, OrderItemDoc, LogisticsDeliveryDoc, InventoryDoc};
